feat: Refactor Dashboard and related components for improved structure and functionality
- Updated ConfidenceBadge to round confidence directly instead of converting to percentage.
- Refactored Dashboard component to utilize new subcomponents: WatchlistPanel, ChartPanel, TradeSetupPanel, and OpenPaperTradesPanel.
- Implemented DashboardController to fetch dashboard data including candles, signals, and news.
- Created PaperTradesController for managing paper trades with endpoints for creating and retrieving trades.
- Added DashboardResponse DTO for structured response data.
- Introduced new components: ChartPanel, NewsPanel, OpenPaperTradesPanel, TradeSetupPanel, and WatchlistPanel for better modularity.
- Enhanced TradeSetupPanel to calculate and display risk/reward ratios and estimated risks/rewards based on user input.
- Improved error handling and logging across controllers.

+++ API/Models/ScanResult.cs
@@ -31,9 +31,9 @@ public class ScanResult
     public bool HasNews { get; set; }
 
     /// <summary>
-    /// Confidence score from signal (0.0 to 1.0).
+    /// Confidence score (0-100).
     /// </summary>
-    public double Confidence { get; set; }
+    public int Confidence { get; set; }
 
     /// <summary>
     /// List of reasons why this stock was scored this way (max 3 for MVP).
